@@ -3,24 +3,21 @@ plugins {
 }
 
 android {
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.michtech.pointofSale"
         minSdk 24
-        targetSdk 33
+        targetSdk 34
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     packagingOptions {
-        exclude 'META-INF/DEPENDENCIES'
-        exclude 'META-INF/NOTICE'
-        exclude 'META-INF/NOTICE.txt'
-        exclude 'META-INF/LICENSE'
-        exclude 'META-INF/LICENSE.txt'
-        exclude 'META-INF/INDEX.LIST'
+        resources {
+            excludes += ['META-INF/DEPENDENCIES', 'META-INF/NOTICE', 'META-INF/NOTICE.txt', 'META-INF/LICENSE', 'META-INF/LICENSE.txt', 'META-INF/INDEX.LIST']
+        }
     }
     buildTypes {
         release {
@@ -29,33 +26,35 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
+    namespace 'com.michtech.pointofSale'
 }
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
+    implementation(platform("org.jetbrains.kotlin:kotlin-bom:1.8.0"))
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.11.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
     implementation 'com.yanzhenjie:album:2.1.3'
-    implementation 'com.google.android.gms:play-services-cast-framework:21.2.0'
-    implementation 'com.github.bumptech.glide:glide:4.13.2'
+    implementation 'com.google.android.gms:play-services-cast-framework:21.4.0'
+    implementation 'com.github.bumptech.glide:glide:4.16.0'
     implementation 'de.hdodenhof:circleimageview:3.1.0'
     implementation 'com.google.ar.sceneform:filament-android:1.17.1'
     implementation 'com.github.castorflex.smoothprogressbar:library-circular:1.3.0'
     implementation 'com.esotericsoftware:kryonet:2.22.0-RC1'
-    implementation "androidx.core:core:1.9.0"
+    implementation "androidx.core:core:1.12.0"
     implementation 'androidx.core:core-google-shortcuts:1.1.0'
     implementation 'com.github.AnyChart:AnyChart-Android:1.1.2'
     implementation 'com.github.adroitandroid:Near:v2.0'
     implementation 'pub.devrel:easypermissions:3.0.0'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0-alpha'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.13.2'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.16.0'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
