@@ -4,8 +4,6 @@ import android.annotation.SuppressLint;
 import android.app.DatePickerDialog;
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.DatePicker;
 import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.PopupMenu;
@@ -14,19 +12,16 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.michtech.pointofSale.R;
-import com.michtech.pointofSale.Ui.more.employee.AddEmployee;
 import com.michtech.pointofSale.adapter.HistoryAdapter;
 import com.michtech.pointofSale.database.DatabaseManager;
 import com.michtech.pointofSale.database.DbHelper;
 import com.michtech.pointofSale.pojo.PojoHistory;
 
-import java.lang.reflect.Method;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
-import java.util.function.Function;
 
 public class History extends AppCompatActivity {
     ListView listView;
@@ -52,57 +47,44 @@ public class History extends AppCompatActivity {
         listShowDay(getDate());
         displayList();
 
-        Back.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                finish();
+        Back.setOnClickListener(view -> finish());
+        onDateSetListener = (datePicker, year, month, day) -> {
+            month = month+1;
+            String dDay = Integer.toString(day);
+            if(day<10){
+                dDay = "0"+day;
             }
-        });
-        onDateSetListener = new DatePickerDialog.OnDateSetListener() {
-            @Override
-            public void onDateSet(DatePicker datePicker, int year, int month, int day) {
-                month = month+1;
-                String dDay = Integer.toString(day);
-                if(day<10){
-                    dDay = "0"+day;
-                }
-                if(month<10){
-                    date = year+"-0"+month+"-"+dDay;
-                }else{
-                    date = year+"-"+month+"-"+dDay;
-                }
-                System.out.println(date);
-                listShowDay(date);
-                displayList();
+            if(month<10){
+                date = year+"-0"+month+"-"+dDay;
+            }else{
+                date = year+"-"+month+"-"+dDay;
             }
+            System.out.println(date);
+            listShowDay(date);
+            displayList();
         };
-        Menu.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                PopupMenu popupMenu = new PopupMenu(History.this, Menu);
-                popupMenu.inflate(R.menu.history_menu);
-                popupMenu.show();
-                popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-                    @SuppressLint("NonConstantResourceId")
-                    @Override
-                    public boolean onMenuItemClick(MenuItem menuItem) {
-                        switch(menuItem.getItemId()){
-                            case R.id.today:
-                                listShowDay(getDate());
-                                displayList();
-                                break;
-                            case R.id.selectDay:
-                                DateDialog();
-                                break;
-                            case R.id.All:
-                                listShowAll();
-                                displayList();
-                                break;
+        Menu.setOnClickListener(view -> {
+            PopupMenu popupMenu = new PopupMenu(History.this, Menu);
+            popupMenu.inflate(R.menu.history_menu);
+            popupMenu.show();
+            popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                @SuppressLint("NonConstantResourceId")
+                @Override
+                public boolean onMenuItemClick(MenuItem menuItem) {
+                    switch (menuItem.getItemId()) {
+                        case R.id.today -> {
+                            listShowDay(getDate());
+                            displayList();
                         }
-                        return false;
+                        case R.id.selectDay -> DateDialog();
+                        case R.id.All -> {
+                            listShowAll();
+                            displayList();
+                        }
                     }
-                });
-            }
+                    return false;
+                }
+            });
         });
 
     }
@@ -129,8 +111,7 @@ public class History extends AppCompatActivity {
     @NonNull
     private String getDate() {
         @SuppressLint("SimpleDateFormat") SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        String currentDateTime = sdf.format(new Date());
-        return currentDateTime;
+        return sdf.format(new Date());
     }
     private void DateDialog(){
         Calendar calendar = Calendar.getInstance();
